rectangle filling, tile drawing changes, time changes

+++ src/Camera.cpp
@@ -5,8 +5,8 @@
 //Camera constructor
 Camera::Camera(int x, int y, int const* CAMERA_MAX_X, int const* CAMERA_MAX_Y)
 {
-	this->x = x;
-	this->y = y;
+	this->x = (double)x;
+	this->y = (double)y;
 	this->CAMERA_MAX_X = CAMERA_MAX_X;
 	this->CAMERA_MAX_Y = CAMERA_MAX_Y;
 }
@@ -16,16 +16,17 @@ Camera::~Camera()
 {
 }
 
-int Camera::getCameraX()
+double Camera::getCameraX()
 {
 	return x;
 }
-int Camera::getCameraY()
+double Camera::getCameraY()
 {
 	return y;
 }
 
-bool Camera::setCameraPos(int x, int y)
+//Sets the cameras position
+bool Camera::setCameraPos(double x, double y)
 {
 	//Check for invalid positions
 	if (x<0 || x>*CAMERA_MAX_X)
