+++ src/Camera.h
@@ -9,14 +9,14 @@ public:
 
 	Camera(int x, int y, int const* CAMERA_MAX_X, int const* CAMERA_MAX_Y);
 	~Camera();
-	int getCameraX();
-	int getCameraY();
-	bool setCameraPos(int x, int y);
+	double getCameraX();
+	double getCameraY();
+	bool setCameraPos(double x, double y);
 
 private:
 
-	int x;
-	int y;
+	double x;
+	double y;
 	int const* CAMERA_MAX_X;
 	int const* CAMERA_MAX_Y;
